move 'pickBy' tests to separate file
